Every article-backed draft was silently losing its source
Rob hit "No integration article for BNB Chain yet" on a row whose badge
said "X article". Both read the same value, so the badge was right and
the draft call was wrong.

Postgres bigint comes back from the driver as a STRING. The driver
cannot know the value is inside JS's safe integer range, so it refuses
to guess. Every id on every shelf is a bigint identity, so articleId
reached /api/generate as "30" rather than 30, and the guard there was
Number.isFinite(articleId) — false for a string. The id was nulled on
the way in.

TypeScript could not catch it. The interfaces all say `number` and the
driver hands back a string at runtime; nothing checks the two agree.

The chain pillar is only where it became VISIBLE, because chain drafts
now fail loudly rather than drafting unsourced. docPageId and videoId
have the identical shape and no such guard, so Dev Doc and Short-Form
Video drafts have been quietly falling through to the pillar-only
prompt for as long as those shelves have existed — the drafter inventing
plausible detail with the actual page and transcript sitting unread.
Same for article-backed Product Posts and Thought Leadership. That is
the bug Rob originally reported, and it was never only about chains.

Fixed at both ends. The shelf queries cast id::int so the TS types are
honest about what arrives, and the route coerces instead of validating,
so a caller cannot silently null an id again.

+++ lib/articles.ts
@@ -137,7 +137,7 @@ export async function getArticleShelf(
         AND (${opts.since}::timestamptz IS NULL OR p.created_at >= ${opts.since}::timestamptz)
     )
     SELECT
-      a.id                                   AS "articleId",
+      a.id::int                              AS "articleId",
       a.slug, a.title, a.dek, a.author,
       to_char(a.published_on, 'YYYY-MM-DD')  AS "publishedOn",
       a.canonical_url                        AS "canonicalUrl",
@@ -175,7 +175,7 @@ export async function getArticleShelf(
   `;
 
   const useRows = await sql<ArticleUse & { articleId: number | null }>`
-    SELECT a.id AS "articleId", p.id, p.url, p.created_at AS "createdAt",
+    SELECT a.id::int AS "articleId", p.id, p.url, p.created_at AS "createdAt",
            EXTRACT(DAY FROM now() - p.created_at)::int AS "daysAgo",
            p.text, p.media_type AS "mediaType", s.impressions,
            (a.anchor_post_id = p.id) AS "isAnchor"
